my_project/requirements.txt
torch==2.2.2
torchvision==0.17.2
fastapi==0.110.0
uvicorn==0.30.0
pillow==10.3.0
numpy<2
python-multipart
matplotlib